requirements.txt
pytest==9.0.2
allure-pytest==2.15.3
selenium==4.41.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
requests==2.32.5